rag_system/requirements.txt
langchain>=0.3.13
langchain-community>=0.0.27
chromadb>=0.4.24
sentence-transformers>=2.5.1
python-dotenv>=1.0.1 